+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 ## 1.1.0
-* New feature: EmojiPickerExtendedUtils provide access to recent emojis, search emoji and adding emoji to recently-used list
+* New feature: EmojiPickerUtils provide access to recent emojis, search emoji and adding emoji to recently-used list
 
 ## 1.0.8